--- cxx/allgather_int.cc
@@ -1,20 +1,0 @@
-/*
- * Copyright (c) 2019 Cisco Systems, Inc.  All rights reserved.
- */
-
-#include "froozle_config.h"
-
-#include <stdio.h>
-
-#include "mpi.hpp"
-
-
-int MPI_Allgather(const void *sendbuf, int sendcount,
-                  MPI_Datatype sendtype,
-                  void *recvbuf, int recvcount,
-                  MPI_Datatype recvtype, MPI_Comm comm)
-{
-    printf("This is C++ %s (with int params)\n", __func__);
-
-    return MPI_SUCCESS;
-}
